Simplify and correct SoftMaxNode::propagate.
Removed SoftMaxNode::get_denominator and relevant C++ tests. Opted to
not implement a method to check if prior_denominator == new_denominator
(which could reduce the # of updates) since this edge case is unlikely
in practical use cases. Opted to continue to cache updated exp() in
std::unordered_map (as opposed to recording the indices that have been
updated and recompute the value at these indices fully with an
additional exp() call) because there was a bug that occurred when an
update was called more than once on the same index. This choice
increases memory usage but should result in faster computations.